backend/requirements.txt
openai-whisper
cohere
Flask
Flask-CORS
pyannote.audio
python-dotenv
pydub
numpy
cv2
flask
opencv-python